Move functionality into new commands
