+++ src/cli/Commands/LangServer.cs
@@ -4,7 +4,11 @@ using Spectre.Console;
 
 namespace Noa.Cli.Commands;
 
-public sealed class LangServer
+public sealed class LangServer(
+    IAnsiConsole console,
+    bool stdio,
+    FileInfo? logFile,
+    LogLevel logLevel)
 {
     public static Command CreateCommand(IAnsiConsole console)
     {
@@ -35,6 +39,37 @@ public sealed class LangServer
         command.Add(logOption);
         command.Add(logLevelOption);
 
+        command.SetAction((ctx, _) =>
+            new LangServer(
+                console,
+                ctx.GetValue(stdioOption),
+                ctx.GetValue(logOption),
+                ctx.GetValue(logLevelOption))
+            .Execute());
+
         return command;
+    }
+
+    private async Task<int> Execute()
+    {
+        if (!stdio)
+        {
+            console.WriteLine(
+                "Language server does not support non-stdio transport kinds. " +
+                "Specify --stdio to launch the server through stdio.",
+                Color.Red);
+            return 1;
+        }
+        
+        try
+        {
+            await NoaLanguageServer.RunAsync(logFile?.FullName, logLevel);
+        }
+        catch (TaskCanceledException) {}
+        catch (OperationCanceledException) {}
+
+        console.WriteLine("Language server died a peaceful death.");
+
+        return 0;
     }
 }
