--- src/cli/LangServer.cs
@@ -1,36 +1,0 @@
-using Cocona;
-using Noa.LangServer;
-using Spectre.Console;
-
-namespace Noa.Cli;
-
-public sealed class LangServer(IAnsiConsole console)
-{
-    [Command("lang-server", Description = "Starts the Noa language server")]
-    public async Task<int> Start(
-        [Option("stdio", Description = "Specifies that the language server should communicate across standard IO")]
-        bool stdio,
-        [Option("log", Description = "A path to a file to which logs will be written.")]
-        string? logFilePath,
-        [Option("log-level", Description = "The level of messages to log from the language server.")]
-        LogLevel logLevel = LogLevel.Info)
-    {
-        if (!stdio)
-        {
-            console.MarkupLine("[red]Language server does not support non-stdio transport kinds. " +
-                               "Specify [/][white]--stdio[/][red] to launch the server through stdio.[/]");
-            return 1;
-        }
-        
-        try
-        {
-            await NoaLanguageServer.RunAsync(logFilePath, logLevel);
-        }
-        catch (TaskCanceledException) {}
-        catch (OperationCanceledException) {}
-
-        console.WriteLine("Language server died a peaceful death.");
-
-        return 0;
-    }
-}
